+++ python/cuspatial/cuspatial/_lib/interpolate.pyx
@@ -1,13 +1,15 @@
 # Copyright (c) 2019-2020, NVIDIA CORPORATION.
 
-# cython: profile=False
-# distutils: language = c++
-# cython: embedsignature = True
-# cython: language_level = 3
-
 from cudf._lib.table cimport table, Table
-from cudf._lib.move cimport move, unique_ptr
 from cudf._lib.column cimport column, Column
+
+from cuspatial._lib.cpp.interpolate cimport (
+    cubicspline_interpolate as cpp_cubicspline_interpolate,
+    cubicspline_coefficients as cpp_cubicspline_coefficients,
+)
+
+from libcpp.utility cimport move
+from libcpp.memory cimport unique_ptr
 
 cpdef cubicspline_coefficients(
     Column t,